+++ lib/src/maintenance.g.dart
@@ -15,25 +15,34 @@ Maintenance _$MaintenanceFromJson(Map<String, dynamic> json) => new Maintenance(
     missingReadme: json['missingReadme'] as bool,
     missingAnalysisOptions: json['missingAnalysisOptions'] as bool,
     oldAnalysisOptions: json['oldAnalysisOptions'] as bool,
+    strongModeEnabled: json['strongModeEnabled'] as bool,
     isExperimentalVersion: json['isExperimentalVersion'] as bool,
     isPreReleaseVersion: json['isPreReleaseVersion'] as bool,
-    toolIssueCount: json['toolIssueCount'] as int);
+    errorCount: json['errorCount'] as int,
+    warningCount: json['warningCount'] as int,
+    hintCount: json['hintCount'] as int);
 
 abstract class _$MaintenanceSerializerMixin {
   bool get missingChangelog;
   bool get missingReadme;
   bool get missingAnalysisOptions;
   bool get oldAnalysisOptions;
+  bool get strongModeEnabled;
   bool get isExperimentalVersion;
   bool get isPreReleaseVersion;
-  int get toolIssueCount;
+  int get errorCount;
+  int get warningCount;
+  int get hintCount;
   Map<String, dynamic> toJson() => <String, dynamic>{
         'missingChangelog': missingChangelog,
         'missingReadme': missingReadme,
         'missingAnalysisOptions': missingAnalysisOptions,
         'oldAnalysisOptions': oldAnalysisOptions,
+        'strongModeEnabled': strongModeEnabled,
         'isExperimentalVersion': isExperimentalVersion,
         'isPreReleaseVersion': isPreReleaseVersion,
-        'toolIssueCount': toolIssueCount
+        'errorCount': errorCount,
+        'warningCount': warningCount,
+        'hintCount': hintCount
       };
 }